4.21 finished

+++ part4/bloglist/controllers/blogs.js
@@ -25,8 +25,21 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(req.params.id)
+    res.status(204).end()
+  } else {
+    return res.status(401).json({ error: 'Unauthorized' })
+  }
+
+  // await Blog.findByIdAndDelete(req.params.id)
 })
 
 blogsRouter.put('/:id', async (req, res) => {
